requirements.txt
attrs==19.3.0
jupyterlab==3.0.7
matplotlib==3.3.4
numpy==1.19.5
OpenNMT-Py==1.2.0
pandas==1.1.5
pint==0.16.1
python-levenshtein==0.12.2
quantulum3==0.7.6
seaborn==0.11.1
textdistance==4.1.5
paragraph2actions @ git+https://github.com/rxn4chemistry/paragraph2actions
